src/housing_streamlit/requirements.txt
numpy
pandas
scikit-learn
streamlit
matplotlib
seaborn
plotly-express
